Moved grid and cell components to own files
