+++ .gitignore
@@ -1,5 +1,6 @@
 .idea/
 build/
+**/__pycache__/
 *.out
 *.so
 *.dll
