+++ c/src/grid.c
@@ -1,0 +1,11 @@
+#include <grid.h>
+#include "quadtree/quadtree.h"
+
+ant_grid_t * new_grid()
+{
+
+	ant_grid_t * grid = calloc(1, sizeof(ant_grid_t));
+	grid->tree = libant_quadtree_create();
+	return grid;
+
+}
